catfile.c: Add --batch-command mode
Add new flag --batch-command that  it will accept commands and arguments
from stdin, similar to git-update-ref --stdin.

At GitLab, we use a pair of long running cat-file processes when
accessing object content. One for iterating over object metadata with
--batch-check, and the other to grab object contents with --batch.

However, if we had --batch-command, we could get rid of the
--batch-check process altogether, and just have one process where we
can flip between getting object info, and getting object contents.
This means we can get rid of roughly half of long lived git cat-file
processes. This can lead to huge savings since on a given server there
could be hundreds of git cat-file processes running.

git cat-file --batch-command

would enter a command mode whereby the user can interactively enter in
commands and their associated arguments:

<command> [arg1] [arg2] NL

This patch adds the plumbing to add commands in the future that can plug
into this command mode.

This patch also adds the following commands:

contents <rev> NL
info <rev> NL

The contents command takes a <rev> argument and prints out the object
contents.

The info command takes a <rev> argument and prints out the object
metadata.

In addition, we need a set of commands that allow transaction semantics.

When a separate process A is connected to a git cat-file process B and is
interactively writing to and reading from it in --buffer mode, A needs
to be able to know when the buffer is flushed to stdout. Currently, from
A's perspective, the only way is to either 1. exit B's process or
2. send an invalid object to stdin. 1. is not ideal from a performance
perspective as it will require spawning a new cat-file process each
time, and 2. is hacky and not a good long term solution.

With the following commands, process A can begin a "transaction" and
send a list of object names over stdin. When "get contents" or "get info"
is received, this list of object names will get fed into
batch_one_object() to retrieve either info or contents. Finally an
fflush() will be called.

begin NL
get contents NL
get info NL

These can be used in the following way:

begin
<sha1>
<sha1>
<sha1>
<sha1>
get info

begin
<sha1>
<sha1>
<sha1>
<sha1>
get contents

Helped-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>
Signed-off-by: John Cai <johncai86@gmail.com>