setup: add_prefix, prefix_path to take a repository argument
To reduce the global state due to the_repository, modify add_prefix to
take a repository argument. This patch ends up expanding the use of
the_repository, but that's because prefix_path was calling it
eventually.

This change hence allows us to get rid of the indirect usage of
the_repository more easily in the future.

Signed-off-by: John Cai <johncai86@gmail.com>