setup: verify_filename to take a repository argument
In order to reduce the prevalence of global state due to the
the_repository variable, allow verify_filename to take a repository
argument.

This ends up expanding the use of the_repository in the grep, reset, and
rev-parse builtins, but verify_filename() was calling the_repository
under the hood, so this change allows the removal of the_repository from
the whole call chain more easily in the future.

Signed-off-by: John Cai <johncai86@gmail.com>