check-ref-format: stop using the_repository
Use the repo passed in through the builtin instead of relying on the
the_repository global variable.

Also, change the flags for the builtin for check-ref-format to add
RUN_SETUP_GENTLY because using check-ref-format with --branch, requires
a repository to be present.

We also no longer need to BUG when prefix is non-null.

Signed-off-by: John Cai <johncai86@gmail.com>